Break Time on the anvil of proptest (#1610)
Our handling of the various datetime types is extremely dubious. In
particular, we `impl From` for these types to allow a conversion from
their "raw repr", the underlying integer type. This was justified on the
idea that this is how Postgres reasons about them anyways, and at worst
it was an application bug. However, there are two problems:

- The conversions are not infallible, as `From` requires[^1]! At various
points of the conversions, they may `ereport` (which we convert to a
`panic!`, naturally).
- For many possible integer values, they return **non-equal** values
when passed through the native Postgres (de)serialization functions and
converted back to the original type, which we should not accept as a
reasonable outcome. It permits data that flows through pgrx extensions
to become corrupt or invalid by the time it reaches the database,
causing errors when read out again. In this sense, it is another way the
conversion is fallible, except the realization of the error may become
delayed for hours or weeks after the data is actually committed to the
database, as it is only upon (de)serialization that it is guaranteed to
error.

Accordingly, this PR breaks the `From` impls for Time, TimeTz, and
Timestamp, for their integer representations. The implementations are
replaced by convenience functions that _do_ implement a possible valid
interpretation of conversion from the integer types, but are lossy, and
thus inappropriate for `impl From`. Conversion to-and-from these integer
types is no longer lossless, and thus it is not "perfect".

The bounds selected for the conversion functions reflect the bounds that
Postgres implements for these types.

Fixes https://github.com/pgcentralfoundation/pgrx/issues/1525

[^1]: https://doc.rust-lang.org/std/convert/trait.From.html