Add a docker image for CI
A basic image with all the requirements to build a PGRX project
in a CI job. This saves a lot of CI minutes, especially the
`cargo pgrx init` step which is quite long.

By default, the image is based on the latest version available
but it can be overriden with `--build-arg PGRX_VERSION=0.11.3`