Support this:
```shell
$ cargo pgrx test pg14 --runas testuser --pgdata=/tmp/pgrxtest
```

This adds two new arguments to `cargo pgrx test`: `--runas <username>` and `--pgdata </writable/path>`

Together, these will initidb, createdb, dropdb, and run postgres as the "runas" user.

Extension artifacts are copied with `cargo pgrx install --sudo`, which will also require the current user be able to become root.

In general, these options are designed for the unusual situation where one needs to be "root" in order to run extension tests, but, of course, Postgres is not allowed to be root.