Allow relocatable extensions (#1653)
Previously, relocatable extensions were disallowed by a check in the
code for the install command. This pull request removes the check, and,
more importantly, changes the behavior in `PgrxSql::schema_alias_of()`
so that it simply never schema-qualifies based on the control file. This
prevents relocatable extensions from trying to schema-qualify with
`@extname@`, which did not get substituted out by Postgres.

This resolves #1626