Do not default `to_entity_graph_tokens` (#1676)
Not providing an impl for this fn would result in nonsensical behavior,
and fortunately nothing does.

There are defaults for the other part of the CodeEnrichment scheme
(`ToRustTokens`) but I am not going to address them right now. The items
in question do expand to Rust tokens, just not under the "purview" of
pgrx-sql-entity-graph.