build(nix): add nix flake with `cargo-pgrx` package
This exposes the `cargo pgrx` binary as a flake package called
`cargo-pgrx`. It can be built using `nix build .#cargo-pgrx`.

The systems specified in this flake were chosen from the project's
README of tested systems. The flake inputs were chosen to minimize the
maintenance burden.

Since the build runs tests in order to succeed, but the tests are
performed at build time for nix, the `test_parse_managed_postmasters`
test has been skipped. Since pgrx already has a test suite that runs
outside of nix, this should be a good compromise to allow the build to
work, but also keep the majority of the tests to ensure that nix users
are aware when they are broken. In addition to this, the `PGRX_HOME`
variable must be set for tests, so it is assigned to a tmpdir before
running the tests.