Deprecate `variadic!` support (#1713)
This removes `variadic!` support. The macro itself remains, but it does
nothing now, and has a deprecation note. There are no uses of it in this
codebase, the ZomboDB codebase, or other pgrx-using codebases I have
seen. `VariadicArray` completely surpasses it in utility. I want to
remove, ideally, all these ad-hoc type macros, but they may need to be
replaced in functionality. This one, however, is already dead.