suppress some warnings in the generated code
User code with `#[pg_extern(...)]` causes `cargo clippy` to show this warning:

```
error: used binding `_fcinfo` which is prefixed with an underscore. A leading underscore signals that a binding will not be used
```

This PR adds a few more anti-warning attributes. Also this PR makes a minor optimization by removing a Vec realloc by simply taking the existing one, instead of cloning and clearing it.