Make unboxing args sound with `ArgAbi` (#1731)
This makes moving arguments into pgrx's `pg_extern` functions sound, by
using a new trait, `ArgAbi`, to describe this behavior. The supporting
`Arg` and `Args` types allow effectively treating the arguments of the
function like an iterator, but the behavior is much more `unsafe`.

## The victories
- A `#[pg_extern] fn` must be sound in its arguments and return types.
- A large amount of our code generation for argument unboxing is now
driven **entirely by type inference**, which simplifies the code
analysis we must do.

## The collateral damage
- This kills a ton of our aggregate examples because they were relying
on `'static` lifetimes and `#[pg_aggregate]` does not yet understand how
to pass through lifetimes correctly.
- Arguments borrowed for `'static` now fail to compile, even in a
hypothetical case it might be okay. This is because pgrx attempts to
translate an argument from a lifetime-bound `FcInfo` into the `'static`
lifetime, which is correctly rejected by Rust's type system.
- Returning borrows with the `'static` lifetime also fails, for similar
reasons. This affects even hypothetical cases it would be okay (e.g. if
the return was *actually* a `&'static str`). Types that are inherently
`'static` like `String` are still fine.

## Future work
- We're going to bump into a lot of bad ergonomics very shortly as
people test the new bounds, and will have to hammer out new API to make
the desired patterns easy to write.
- This does not retire `FromDatum` and thus doesn't hammer the nail in
the coffin on the unsoundness in the numerous SPI functions that use
`FromDatum`.
- I have not yet begun to cull all the now-unnecessary code that tries
to parse out types, but given that arguments are now driven from
inference? There's a *lot* of it.