Add bindings for 'jit/jit.h', 'utils/resowner.h' and 'utils/resowner_private.h'
These headers are required when writing a JIT provider for PostgreSQL.
Most things in jit.h are required and from resowner_private.h the
ResourceOwner*JIT() functions are necessary.

Additionally jit.h contains the struct used when measuring how much time
is spent doing just-in-time compilation which is useful when working
with EXPLAIN ANALYZE outputs.

The resowner_private.h header will be removed in PostgreSQL 17 and all
functions will be moved to resowner.h so we add bindings for it too.