Update versions even if no specifier (#1803)
Found a bug after doing some refactoring to improve how this tool works
in general, making it more resilient to whatever the hell we do to our
repo. This should allow us to have packages with different names than
the paths to them, allowing arbitrary nesting and relocation, and also
applies a less ad-hoc rule regarding "what packages are we updating".

Unblocks #1802.