Add an example of a WAL decoder
This example shows how to build a basic Change Data Capture (CDC)
mecanism using the Postgres Logical Decoding capabilities. The changes
occuring on the database are serialized into JSON and pushed to a
queue (a "logical replication slot") where they can be consumed by
remote clients.

A Postgres CDC extension can be a used in various purposes:

* Adhoc replication systems (e.g. Postgres => SQL Server )
* External commit-log for a distributed system (such as Khafka)
* Advanced Monitoring ( e.g. Prometheus/Loki )

This example tries to find the right tradeoff between simplicity and
usefulness. Currently it has strong limitations but they should be
easy to overcome.

Rust really shine in this example compared to similar implementations
in C (see the wal2json extension). The Serde crate provides JSON
serialization out of the box, whereas all the C implementations are
forced to rewrite their JSON formater.