Added into_iter for anyarray (#1851)
Fixes #1832.

The implementation isn't what I want, but it adds `into_iter` for
`&AnyArray` and I hope unblocking thomcc. Later it would be better to
get rid of `AnyArray` in its current state and use `Array<AnyElement>`
instead, I will investigate that possibility. But for now that is how it
is and without any breaking change.