`pg_sys::Oid::from_datum()` should cast to `u32`
As implemented in Postgres, the OID version of a `Datum`
is a cast the Datum `uintptr_t` to a `u32`, not trying to
bounds check the pointer value and coerece it into a u32.

Adds a test for this has been added.  It would fail because
`pg_sys::Oid::try_from()` would return an Err on sufficiently large OID
values, which we'd coerece into a None, causing our (at least) function
argument handling to panic, indicating that it received a NULL
post-FromDatum when it shouldn't.

Related, the `PgRelation::from_datum()` implementation is actually a
REGCLASS behind the scenes, which is an OID, and it was similarly
bugged.  This is kinda hard to write a test for as you'd need a relation
of some kind in the schema with an OID over i32::MAX, and that's really
tough to materialize without mucking around with the `pg_resetwal` tool,
which is way outside the scope of what our test suite can handle.