Instruct `rustfmt` to use `--edition 2024`
This is necessary in order for `rustfmt` to understand our
(relatively-new) usages of rust `c"C Strings"`.

However, the word `gen` has been promoted to a keyword in 2024 edition,
which borks `rustfmt` in a different way.  Fixing this requires that we
have a version of bindgen with
https://github.com/rust-lang/rust-bindgen/pull/3100, which, of course,
has not been released yet.  So I don't know where to go next.