teach `#[pg_cast]` to support 3-argument CAST functions
CAST functions can have 1 argument or 3 arguments where the latter is
the source type plus the type modifier ("typmod") and explicit context
for the destination type.

This changes `#[pg_cast]` to blindly support any number of arguments on
the function to which it's attach and then Postgres will sort it out
during DDL playback.