give exported `__pgrx_internals_fn_*` functions unique names
It's entirely possible for a pgrx extension to have functions like this:

```rust
mod a {
   #[pg_extern]
   fn foo() { }
}

mod b {
   #[pg_extern]
   fn foo() { }
}
```

These are two distinct functions in Rust, and should be in the SQL
schema, but the naming of the sql entity graph `__pgrx_internals_fn_*`
function isn't unique enough.

We could also include the schema name there, but I don't think that's
quite enough, so instead we just append a UUIDv4 value.