Remove `impl AsRef<Path>` from cargo-pgrx/src/schema.rs
A pattern that was briefly in vogue some years ago, when impl Trait args
were first introduced, was to make everything an impl Trait arg.
This can be undesirable because it weakens type inference and spreads
whatever indirection is used throughout the codebase, demanding repeat
invocations of `.as_ref()` or whatever.

Clear the codebase of it, starting with schema.rs in cargo-pgrx